Tried out MalbehavD v1 as a real world data.
TLDR, it did not work.

The author actually did mention that it underwent process "to remove all inconsistencies/noise, making it ready to be used for evaluating the performance of deep learning models."

This implies that there might be steps that are different to that of Oliveira's dataset and that the MalbehavD, while can work on ML models, is primarily designed for DL. That being the case, let's just consider the 1st reasoning as it is the most sound one.